requirements.txt
aiohappyeyeballs==2.6.1
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.5
gunicorn==23.0.0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.4
sqlparse==0.5.3
whitenoise==6.11.0
psycopg[binary]>=3.0.0
django-storages==1.14.0
boto3==1.28.0
pillow==12.0.0
